Potenziamento controlli eliminazione contratto

+++ modules/contratti/actions.php
@@ -259,15 +259,26 @@ switch (post('op')) {
 
     // eliminazione contratto
     case 'delete':
-        try {
-            $contratto->delete();
 
-            $dbo->query('DELETE FROM co_promemoria WHERE idcontratto='.prepare($id_record));
-            $dbo->query('DELETE FROM co_contratti_tipiintervento WHERE idcontratto='.prepare($id_record));
+        // Fatture o interventi collegati a questo contratto
+        $elementi = $dbo->fetchArray('SELECT 0 AS `codice`, `co_documenti`.`id` AS `id`, `co_documenti`.`numero` AS `numero`, `co_documenti`.`numero_esterno` AS `numero_esterno`,  `co_documenti`.`data`, `co_tipidocumento`.`descrizione` AS `tipo_documento`, `co_tipidocumento`.`dir` AS `dir`  FROM `co_documenti` JOIN `co_tipidocumento` ON `co_tipidocumento`.`id` = `co_documenti`.`idtipodocumento` WHERE `co_documenti`.`id` IN (SELECT `iddocumento` FROM `co_righe_documenti` WHERE `idcontratto` = '.prepare($id_record).')'.'
+        UNION
+        SELECT  `in_interventi`.`codice` AS `codice`, `in_interventi`.`id` AS `id`, 0 AS `numero`, 0 AS `numero_esterno`, `in_interventi`.`data_richiesta` AS `data`, 0 AS `tipo_documento`, 0 AS `dir` FROM `in_interventi` WHERE `in_interventi`.`id_contratto` = '.prepare($id_record).' ORDER BY `data` ');
 
-            flash()->info(tr('Contratto eliminato!'));
-        } catch (InvalidArgumentException $e) {
-            flash()->error(tr('Sono stati utilizzati alcuni serial number nel documento: impossibile procedere!'));
+        if (empty($elementi)) {
+
+            try {
+                $contratto->delete();
+
+                $dbo->query('DELETE FROM co_promemoria WHERE idcontratto='.prepare($id_record));
+                $dbo->query('DELETE FROM co_contratti_tipiintervento WHERE idcontratto='.prepare($id_record));
+                $dbo->query('DELETE FROM my_impianti_contratti WHERE idcontratto='.prepare($id_record));
+
+                flash()->info(tr('Contratto eliminato!'));
+            } catch (InvalidArgumentException $e) {
+                flash()->error(tr('Sono stati utilizzati alcuni serial number nel documento: impossibile procedere!'));
+            }
+
         }
 
         break;
